Update : Qemu handling, Report Save change

+++ app/scripts/qemu_script.py
@@ -5,39 +5,30 @@ import time
 
 class QemuManager:
     def __init__(self) -> None:
-        pass
+        self.qemu_process = None
 
     def start_qemu(self):
-            # Define QEMU command with monitor enabled
-            qemu_command = [
-                r'app\qemu\qemu-system-x86_64.exe',  # Path to QEMU executable
-                '-m',  '8G',
-                '-smp', '4',
-                '-hda', r'app\qemu\kali\kali.qcow2',
-                '-usbdevice', 'tablet',
-                '-name', 'kali',
-                '-nic', 'user,restrict=off,model=virtio,id=vmnic,hostfwd=tcp::60022-:22,hostfwd=tcp::9392-:9392',
-                '-monitor', 'stdio',
-                '-vga',  'vmware',
-                '-loadvm', 'gvm',
-                '-vnc', ':0'
-            ]
+        qemu_command = [
+            r'app\qemu\qemu-system-x86_64.exe',
+            '-m', '8G',
+            '-smp', '4',
+            '-hda', r'app\qemu\kali\kali.qcow2',
+            '-usbdevice', 'tablet',
+            '-name', 'kali',
+            '-nic', 'user,restrict=off,model=virtio,id=vmnic,hostfwd=tcp::60022-:22,hostfwd=tcp::9392-:9392',
+            '-monitor', 'stdio',
+            '-vga', 'vmware',
+            '-loadvm', 'gvm',
+            '-vnc', ':0'
+        ]
 
-            # Start QEMU in a separate thread
-            self.qemu_thread = threading.Thread(target=self._start_qemu_thread, args=(qemu_command,))
-            self.qemu_thread.start()
-
-    def _start_qemu_thread(self, qemu_command):
-        # Start QEMU
         self.qemu_process = subprocess.Popen(qemu_command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-        # Wait for QEMU to boot (you may need to adjust the sleep duration)
 
     def terminate_qemu(self):
-        # Terminate QEMU process
-        self.qemu_process.terminate()
-
-        # Wait for the QEMU thread to join (wait for it to finish)
-        self.qemu_thread.join()
+        print(self.qemu_process)
+        if self.qemu_process:
+            self.qemu_process.terminate()
+            self.qemu_process.wait()
 
     def prep_kali(self):
 
@@ -65,10 +56,6 @@ class QemuManager:
 
         # Wait for user input before terminating QEMU
         #input("Press Enter to terminate QEMU...")
-
-    def terminate_qemu(self):
-        #Terminate QEMU process
-        self.qemu_process.terminate()
 
 class QemuSSHManager:
     def __init__(self):
